完成码匠37～38的内容，利用Postman测试，实现了后端接收POST格式数据。重写了异常捕捉的判断，调整了表结构和对应方法，修复若干BUG

+++ src/main/java/com/sapphire/demo/model/Question.java
@@ -8,7 +8,7 @@ public class Question {
      *
      * @mbg.generated
      */
-    private Integer id;
+    private Long id;
 
     /**
      *
@@ -53,7 +53,7 @@ public class Question {
      *
      * @mbg.generated
      */
-    private Integer creator;
+    private Long creator;
 
     /**
      *
@@ -108,7 +108,7 @@ public class Question {
      *
      * @mbg.generated
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -120,7 +120,7 @@ public class Question {
      *
      * @mbg.generated
      */
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -228,7 +228,7 @@ public class Question {
      *
      * @mbg.generated
      */
-    public Integer getCreator() {
+    public Long getCreator() {
         return creator;
     }
 
@@ -240,7 +240,7 @@ public class Question {
      *
      * @mbg.generated
      */
-    public void setCreator(Integer creator) {
+    public void setCreator(Long creator) {
         this.creator = creator;
     }
 
@@ -363,5 +363,4 @@ public class Question {
     public void setDescription(String description) {
         this.description = description == null ? null : description.trim();
     }
-    
 }
